Refactor : crud functions in separate files
