+++ ai_service/api/auth.py
@@ -9,9 +9,12 @@ from jose import JWTError, jwt
 from passlib.context import CryptContext
 from pydantic import BaseModel
 from sqlalchemy.orm import Session
+from ..db import database
+from ai_service.db.crud import (
+    get_user_by_email,
+    create_user,
+)
 
-from ..db import database, prisma_crud
-from ..db import prisma_models
 
 load_dotenv()
 
@@ -59,7 +62,7 @@ def get_password_hash(password):
     return pwd_context.hash(password)
 
 def authenticate_user(db: Session, email: str, password: str):
-    user = prisma_crud.get_user_by_email(db, email)
+    user = get_user_by_email(db, email)
     if not user:
         return False
     if not verify_password(password, user.password):
@@ -90,7 +93,7 @@ async def get_current_user(token: str = Depends(oauth2_scheme), db: Session = De
         token_data = TokenData(email=email)
     except JWTError:
         raise credentials_exception
-    user = prisma_crud.get_user_by_email(db, email=token_data.email)
+    user = get_user_by_email(db, email=token_data.email)
     if user is None:
         raise credentials_exception
     return user
@@ -127,14 +130,14 @@ async def login_for_access_token(form_data: OAuth2PasswordRequestForm = Depends(
 
 @router.post("/register", response_model=User)
 async def register_user(user: UserCreate, db: Session = Depends(database.get_db_prisma)):
-    db_user = prisma_crud.get_user_by_email(db, email=user.email)
+    db_user = get_user_by_email(db, email=user.email)
     if db_user:
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
             detail="Email already registered",
         )
     hashed_password = get_password_hash(user.password)
-    new_user = prisma_crud.create_user(db, user={
+    new_user = create_user(db, user={
         "name": user.name,
         "email": user.email,
         "password": hashed_password,
