+++ ai_service/db/crud/__init__.py
@@ -1,0 +1,11 @@
+from .user import *
+from .dashboard import *
+from .task import *
+from .tasklist import *
+from .tag import *
+from .subtask import *
+from .feedback import *
+from .aitask import *
+from .knowledge import *
+from .conversation import *
+from .prompt import *
