+++ ai_service/routes/tag.py
@@ -1,33 +1,38 @@
 from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
-from ai_service.db import prisma_crud
 from ai_service.db.database import get_db_prisma
 from ai_service.db.schemas import TagCreate, Tag
+from ai_service.db.crud import (
+    create_tag as crud_create_tag,
+    get_tag as crud_get_tag,
+    delete_tag as crud_delete_tag,
+    update_tag as crud_update_tag
+)
 
 router = APIRouter()
 
 @router.post("/tags/", response_model=Tag)
-def create_tag(tag: TagCreate, db: Session = Depends(get_db_prisma)):
-    return prisma_crud.create_tag(db, tag)
+def create_tag_route(tag: TagCreate, db: Session = Depends(get_db_prisma)):
+    return crud_create_tag(db, tag)
 
 @router.get("/tags/{tag_id}", response_model=Tag)
-def get_tag(tag_id: int, db: Session = Depends(get_db_prisma)):
-    db_tag = prisma_crud.get_tag(db, tag_id)
+def get_tag_route(tag_id: int, db: Session = Depends(get_db_prisma)):
+    db_tag = crud_get_tag(db, tag_id)
     if db_tag is None:
         raise HTTPException(status_code=404, detail="Tag not found")
     return db_tag
 
 @router.delete("/tags/{tag_id}", response_model=Tag)
-def delete_tag(tag_id: int, db: Session = Depends(get_db_prisma)):
-    db_tag = prisma_crud.get_tag(db, tag_id)
+def delete_tag_route(tag_id: int, db: Session = Depends(get_db_prisma)):
+    db_tag = crud_get_tag(db, tag_id)
     if db_tag is None:
         raise HTTPException(status_code=404, detail="Tag not found")
-    prisma_crud.delete_tag(db, tag_id)
+    crud_delete_tag(db, tag_id)
     return db_tag
 
 @router.put("/tags/{tag_id}", response_model=Tag)
-def update_tag(tag_id: int, tag: TagCreate, db: Session = Depends(get_db_prisma)):
-    db_tag = prisma_crud.get_tag(db, tag_id)
+def update_tag_route(tag_id: int, tag: TagCreate, db: Session = Depends(get_db_prisma)):
+    db_tag = crud_get_tag(db, tag_id)
     if db_tag is None:
         raise HTTPException(status_code=404, detail="Tag not found")
-    return prisma_crud.update_tag(db, tag_id, tag)
+    return crud_update_tag(db, tag_id, tag)
